Split out push and link handling for josh-cli
This also allows to have one unified #[cfg] directive for the whole link mod
