+++ josh-cli/src/bin/josh.rs
@@ -1,10 +1,10 @@
-#![warn(unused_extern_crates)]
-
 use anyhow::{Context, anyhow};
 use clap::Parser;
 
+#[cfg(feature = "incubating")]
+use josh_cli::commands::link::LinkArgs;
+use josh_cli::commands::push::PushArgs;
 use josh_cli::config::{RemoteConfig, read_remote_config, write_remote_config};
-use josh_core::changes::{PushMode, build_to_push};
 use josh_core::git::{normalize_repo_path, spawn_git_command};
 
 #[derive(Debug, clap::Parser)]
@@ -111,43 +111,6 @@ pub struct FetchArgs {
 }
 
 #[derive(Debug, clap::Parser)]
-pub struct PushArgs {
-    /// Remote name (or URL) to push to (optional, defaults to git's configured remote)
-    ///
-    /// When omitted, behaves like `git push` and uses the current branch's
-    /// configured remote (or a reasonable default such as `origin`).
-    #[arg()]
-    pub remote: Option<String>,
-
-    /// One or more refspecs to push (e.g. main, HEAD:refs/heads/main)
-    ///
-    /// These are positional arguments following the optional remote, matching
-    /// `git push [<repository> [<refspec>...]]` syntax.
-    #[arg()]
-    pub refspecs: Vec<String>,
-
-    /// Force update (non-fast-forward)
-    #[arg(short = 'f', long = "force", action = clap::ArgAction::SetTrue)]
-    pub force: bool,
-
-    /// Atomic push (all-or-nothing if server supports it)
-    #[arg(long = "atomic", action = clap::ArgAction::SetTrue)]
-    pub atomic: bool,
-
-    /// Dry run (don't actually update remote)
-    #[arg(long = "dry-run", action = clap::ArgAction::SetTrue)]
-    pub dry_run: bool,
-
-    /// Use split mode for pushing (defaults to normal mode)
-    #[arg(long = "split", action = clap::ArgAction::SetTrue)]
-    pub split: bool,
-
-    /// Use stack mode for pushing (defaults to normal mode)
-    #[arg(long = "stack", action = clap::ArgAction::SetTrue)]
-    pub stack: bool,
-}
-
-#[derive(Debug, clap::Parser)]
 pub struct RemoteArgs {
     /// Remote subcommand
     #[command(subcommand)]
@@ -184,51 +147,6 @@ pub struct FilterArgs {
     /// Remote name to apply filtering to
     #[arg()]
     pub remote: String,
-}
-
-#[derive(Debug, clap::Parser)]
-#[cfg(feature = "incubating")]
-pub struct LinkArgs {
-    /// Link subcommand
-    #[command(subcommand)]
-    pub command: LinkCommand,
-}
-
-#[cfg(feature = "incubating")]
-#[derive(Debug, clap::Subcommand)]
-pub enum LinkCommand {
-    /// Add a link with optional filter and target branch
-    Add(LinkAddArgs),
-    /// Fetch from existing link files
-    Fetch(LinkFetchArgs),
-}
-
-#[cfg(feature = "incubating")]
-#[derive(Debug, clap::Parser)]
-pub struct LinkAddArgs {
-    /// Path where the link will be mounted
-    #[arg()]
-    pub path: String,
-
-    /// Remote repository URL
-    #[arg()]
-    pub url: String,
-
-    /// Optional filter to apply to the linked repository
-    #[arg()]
-    pub filter: Option<String>,
-
-    /// Target branch to link (defaults to HEAD)
-    #[arg(long = "target")]
-    pub target: Option<String>,
-}
-
-#[cfg(feature = "incubating")]
-#[derive(Debug, clap::Parser)]
-pub struct LinkFetchArgs {
-    /// Optional path to specific .link.josh file (if not provided, fetches all)
-    #[arg()]
-    pub path: Option<String>,
 }
 
 fn main() {
@@ -281,11 +199,11 @@ fn run_command(cli: &Cli) -> anyhow::Result<()> {
         Command::Clone(args) => handle_clone(args, &transaction),
         Command::Fetch(args) => handle_fetch(args, &transaction),
         Command::Pull(args) => handle_pull(args, &transaction),
-        Command::Push(args) => handle_push(args, &transaction),
+        Command::Push(args) => josh_cli::commands::push::handle_push(args, &transaction),
         Command::Remote(args) => handle_remote(args, &transaction),
         Command::Filter(args) => handle_filter(args, &transaction),
         #[cfg(feature = "incubating")]
-        Command::Link(args) => handle_link(args, &transaction),
+        Command::Link(args) => josh_cli::commands::link::handle_link(args, &transaction),
     }
 }
 
@@ -571,382 +489,6 @@ fn handle_fetch(
 
     // Note: fetch doesn't checkout, it just updates the refs
     eprintln!("Fetched from remote: {}", args.remote);
-
-    Ok(())
-}
-
-fn handle_push(args: &PushArgs, transaction: &josh_core::cache::Transaction) -> anyhow::Result<()> {
-    // Read remote configuration from .git/josh/remotes/<name>.josh
-    let repo = transaction.repo();
-    let repo_path = normalize_repo_path(repo.path());
-
-    // Determine which remote to use:
-    // - If a remote was explicitly provided, use it.
-    // - Otherwise, fall back to a reasonable default (currently \"origin\"),
-    //   similar to how `git push` uses the configured upstream when no
-    //   repository argument is given.
-    let remote_name = args.remote.as_deref().unwrap_or("origin");
-
-    let RemoteConfig {
-        url,
-        filter_with_meta,
-        ..
-    } = read_remote_config(&repo_path, remote_name)
-        .with_context(|| format!("Failed to read remote config for '{}'", remote_name))?;
-
-    // Get the wrapped filter (peel away metadata)
-    let filter = filter_with_meta.peel();
-
-    // Get git config for user email
-    let config = repo.config().context("Failed to get git config")?;
-
-    // If no refspecs provided, push the current branch
-    let refspecs = if args.refspecs.is_empty() {
-        // Get the current branch name
-        let head = repo.head().context("Failed to get HEAD")?;
-
-        let current_branch = head
-            .shorthand()
-            .context("Failed to get current branch name")?;
-
-        vec![current_branch.to_string()]
-    } else {
-        args.refspecs.clone()
-    };
-
-    // For each refspec, we need to:
-    // 1. Get the current commit of the local ref
-    // 2. Use Josh API to unapply the filter
-    // 3. Push the unfiltered result to the remote
-
-    for refspec in &refspecs {
-        let (local_ref, remote_ref) = if let Some(colon_pos) = refspec.find(':') {
-            let local = &refspec[..colon_pos];
-            let remote = &refspec[colon_pos + 1..];
-            (local.to_string(), remote.to_string())
-        } else {
-            // If no colon, push local ref to remote with same name
-            (refspec.clone(), refspec.clone())
-        };
-        let remote_ref = remote_ref
-            .strip_prefix("refs/heads/")
-            .unwrap_or(&remote_ref);
-
-        // Get the current commit of the local ref
-        let local_commit = repo
-            .resolve_reference_from_short_name(&local_ref)
-            .with_context(|| format!("Failed to resolve local ref '{}'", local_ref))?
-            .target()
-            .context("Failed to get target of local ref")?;
-
-        // Look up the josh remote reference once and derive both original_target
-        // and old_filtered_oid from it
-        let josh_remote_ref = format!("refs/josh/remotes/{}/{}", remote_name, remote_ref);
-        let (original_target, old_filtered_oid) =
-            if let Ok(remote_reference) = repo.find_reference(&josh_remote_ref) {
-                let josh_remote_oid = remote_reference.target().unwrap_or(git2::Oid::zero());
-
-                // Apply the filter to get the old filtered oid
-                let (filtered_oids, errors) = josh_core::filter_refs(
-                    transaction,
-                    filter,
-                    &[(josh_remote_ref.clone(), josh_remote_oid)],
-                );
-
-                if let Some(error) = errors.into_iter().next() {
-                    return Err(anyhow!("josh filter error: {}", error.1));
-                }
-
-                let old_filtered = if let Some((_, filtered_oid)) = filtered_oids.first() {
-                    *filtered_oid
-                } else {
-                    git2::Oid::zero()
-                };
-
-                (josh_remote_oid, old_filtered)
-            } else {
-                (git2::Oid::zero(), git2::Oid::zero())
-            };
-
-        log::debug!("old_filtered_oid: {:?}", old_filtered_oid);
-        log::debug!("original_target: {:?}", original_target);
-
-        // Set push mode based on the metadata
-        let push_mode = if args.split {
-            PushMode::Split
-        } else if args.stack {
-            PushMode::Stack
-        } else {
-            PushMode::Normal
-        };
-
-        // Get author email from git config
-        let author = config.get_string("user.email").unwrap_or_default();
-
-        let mut changes: Option<Vec<josh_core::Change>> =
-            if push_mode == PushMode::Stack || push_mode == PushMode::Split {
-                Some(vec![])
-            } else {
-                None
-            };
-
-        // Use Josh API to unapply the filter
-        let unfiltered_oid = josh_core::history::unapply_filter(
-            transaction,
-            filter,
-            original_target,
-            old_filtered_oid,
-            local_commit,
-            josh_core::history::OrphansMode::Keep,
-            None,         // reparent_orphans
-            &mut changes, // change_ids
-        )
-        .context("Failed to unapply filter")?;
-
-        log::debug!("unfiltered_oid: {:?}", unfiltered_oid);
-
-        let to_push = build_to_push(
-            transaction.repo(),
-            changes,
-            push_mode,
-            &remote_ref,
-            &author,
-            &remote_ref,
-            unfiltered_oid,
-            original_target,
-        )
-        .context("Failed to build to push")?;
-
-        log::debug!("to_push: {:?}", to_push);
-
-        // Process each entry in to_push (similar to josh-proxy)
-        for (refname, oid, _) in to_push {
-            // Build git push command
-            let mut git_push_args = vec!["push"];
-
-            if args.force || push_mode == PushMode::Split {
-                git_push_args.push("--force");
-            }
-
-            if args.atomic {
-                git_push_args.push("--atomic");
-            }
-
-            if args.dry_run {
-                git_push_args.push("--dry-run");
-            }
-
-            // Determine the target remote URL
-            let target_remote = url.clone();
-
-            // Create refspec: oid:refname
-            let push_refspec = format!("{}:{}", oid, refname);
-
-            git_push_args.push(&target_remote);
-            git_push_args.push(&push_refspec);
-
-            // Use direct spawn so users can see git push progress
-            spawn_git_command(
-                repo.path(),
-                &git_push_args, // Skip "git" since spawn_git_command adds it
-                &[],
-            )
-            .context("git push failed")?;
-
-            eprintln!("Pushed {} to {}/{}", oid, remote_name, refname);
-        }
-    }
-
-    Ok(())
-}
-
-#[cfg(feature = "incubating")]
-fn handle_link(args: &LinkArgs, transaction: &josh_core::cache::Transaction) -> anyhow::Result<()> {
-    match &args.command {
-        LinkCommand::Add(add_args) => handle_link_add(add_args, transaction),
-        LinkCommand::Fetch(fetch_args) => handle_link_fetch(fetch_args, transaction),
-    }
-}
-
-#[cfg(feature = "incubating")]
-fn handle_link_add(
-    args: &LinkAddArgs,
-    transaction: &josh_core::cache::Transaction,
-) -> anyhow::Result<()> {
-    use josh_link::make_signature;
-
-    let repo = transaction.repo();
-
-    // Validate the path (should not be empty and should be a valid path)
-    if args.path.is_empty() {
-        return Err(anyhow!("Path cannot be empty"));
-    }
-
-    // Get the filter (default to ":/" if not provided)
-    let filter = args.filter.as_deref().unwrap_or(":/");
-
-    // Get the target branch (default to "HEAD" if not provided)
-    let target = args.target.as_deref().unwrap_or("HEAD");
-
-    // Use git fetch shell command
-    let output = std::process::Command::new("git")
-        .args(&["fetch", &args.url, target])
-        .output()
-        .context("Failed to execute git fetch")?;
-
-    if !output.status.success() {
-        return Err(anyhow!(
-            "git fetch failed: {}",
-            String::from_utf8_lossy(&output.stderr)
-        ));
-    }
-
-    // Get the commit SHA from FETCH_HEAD
-    let fetch_head = repo
-        .find_reference("FETCH_HEAD")
-        .context("Failed to find FETCH_HEAD")?;
-    let fetch_commit = fetch_head
-        .peel_to_commit()
-        .context("Failed to get FETCH_HEAD commit")?;
-    let fetched_commit = fetch_commit.id();
-
-    // Get the current HEAD commit
-    let head_ref = repo.head().context("Failed to get HEAD")?;
-    let head_commit = head_ref
-        .peel_to_commit()
-        .context("Failed to get HEAD commit")?;
-    let head_tree = head_commit.tree().context("Failed to get HEAD tree")?;
-
-    // Create a new commit with the updated tree
-    let signature = make_signature(&repo)?;
-
-    let result = josh_link::prepare_link_add(
-        transaction,
-        std::path::Path::new(&args.path),
-        &args.url,
-        args.filter.as_deref(),
-        target,
-        fetched_commit,
-        &head_tree,
-    )?
-    .into_commit(transaction, &head_commit, &signature)?;
-
-    // Create the fixed branch name
-    let branch_name = "refs/heads/josh-link";
-
-    // Create or update the branch reference
-    repo.reference(branch_name, result.filtered_commit, true, "josh link add")
-        .with_context(|| format!("Failed to create branch '{}'", branch_name))?;
-
-    let normalized_path = args.path.trim_matches('/');
-    println!(
-        "Added link '{}' with URL '{}', filter '{}', and target '{}'",
-        normalized_path, args.url, filter, target
-    );
-    println!("Created branch: {}", branch_name);
-
-    Ok(())
-}
-
-#[cfg(feature = "incubating")]
-fn handle_link_fetch(
-    args: &LinkFetchArgs,
-    transaction: &josh_core::cache::Transaction,
-) -> anyhow::Result<()> {
-    use josh_link::make_signature;
-
-    let repo = transaction.repo();
-
-    // Get the current HEAD commit
-    let head_ref = repo.head().context("Failed to get HEAD")?;
-    let head_commit = head_ref
-        .peel_to_commit()
-        .context("Failed to get HEAD commit")?;
-    let head_tree = head_commit.tree().context("Failed to get HEAD tree")?;
-
-    let link_files = if let Some(path) = &args.path {
-        // Single path specified - use find_link_files to get all link files, then find the one at the specified path
-        let link_files = josh_core::link::find_link_files(&repo, &head_tree)
-            .context("Failed to find link files")?;
-
-        let link_file = link_files
-            .iter()
-            .find(|(p, _)| p.to_string_lossy() == path.as_str())
-            .map(|(_, lf)| lf.clone())
-            .ok_or_else(|| anyhow!("Link file not found at path '{}'", path))?;
-
-        vec![(std::path::PathBuf::from(path), link_file)]
-    } else {
-        // No path specified - find all .link.josh files in the tree
-        josh_core::link::find_link_files(&repo, &head_tree).context("Failed to find link files")?
-    };
-
-    if link_files.is_empty() {
-        return Err(anyhow!("No .link.josh files found"));
-    }
-
-    println!("Found {} link file(s) to fetch", link_files.len());
-
-    // Fetch from all the link files and collect (path, new_oid) pairs
-    let mut links_to_update = Vec::new();
-    for (path, link_file) in &link_files {
-        println!("Fetching from link at path: {}", path.display());
-
-        // Get remote and branch from metadata
-        let remote = link_file.get_meta("remote").ok_or_else(|| {
-            anyhow!(
-                "Link file missing 'remote' metadata at path '{}'",
-                path.display()
-            )
-        })?;
-        let branch = link_file
-            .get_meta("target")
-            .unwrap_or_else(|| "HEAD".to_string());
-
-        // Use git fetch shell command
-        let output = std::process::Command::new("git")
-            .args(&["fetch", &remote, &branch])
-            .output()
-            .context("Failed to execute git fetch")?;
-
-        if !output.status.success() {
-            return Err(anyhow!(
-                "git fetch failed for path '{}': {}",
-                path.display(),
-                String::from_utf8_lossy(&output.stderr)
-            ));
-        }
-
-        // Get the commit SHA from FETCH_HEAD
-        let fetch_head = repo
-            .find_reference("FETCH_HEAD")
-            .context("Failed to find FETCH_HEAD")?;
-        let fetch_commit = fetch_head
-            .peel_to_commit()
-            .context("Failed to get FETCH_HEAD commit")?;
-        let new_oid = fetch_commit.id();
-
-        links_to_update.push((path.clone(), new_oid));
-    }
-
-    let signature = make_signature(&repo)?;
-    let result = josh_link::update_links(
-        &repo,
-        transaction,
-        &head_commit,
-        links_to_update,
-        &signature,
-    )?;
-
-    // Create the fixed branch name
-    let branch_name = "refs/heads/josh-link";
-
-    // Create or update the branch reference
-    repo.reference(branch_name, result.filtered_commit, true, "josh link fetch")
-        .with_context(|| format!("Failed to create branch '{}'", branch_name))?;
-
-    println!("Updated {} link file(s)", link_files.len());
-    println!("Created branch: {}", branch_name);
 
     Ok(())
 }
