+++ josh-cli/src/commands/mod.rs
@@ -1,1 +1,3 @@
-
+#[cfg(feature = "incubating")]
+pub mod link;
+pub mod push;
